+++ src/main/java/org/gsdistance/grimmsServer/Events/Listeners/EntityDeathEvent.java
@@ -9,35 +9,35 @@ import org.gsdistance.grimmsServer.Stats.ServerStats;
 import org.gsdistance.grimmsServer.Stats.WorldStats;
 
 public class EntityDeathEvent {
-   public EntityDeathEvent() {
-   }
+    public EntityDeathEvent() {
+    }
 
-   public static void Event(org.bukkit.event.entity.EntityDeathEvent event) {
-      if (event.getEntity().getKiller() != null && event.getEntity().getKiller().getType() == EntityType.PLAYER) {
-         PlayerStats playerStats = PlayerStats.getPlayerStats(event.getEntity().getKiller());
-         WorldStats worldStats = WorldStats.getWorldStats(event.getEntity().getWorld());
-         playerStats.changeStat("total_kill_count", 1);
-         playerStats.changeStat("money", (int)Math.round((double)3.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getLesserMoneyMultiplier()));
-         playerStats.changeStat("tPoint", (int)Math.round((double)25.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getMoneyMultiplier()));
-         worldStats.changeStat("wPoint", 40);
-         PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).addExp((double)50.0F);
-         if (event.getEntity().getType() == EntityType.PLAYER) {
-            playerStats.changeStat("money", (int)Math.round((double)12.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getLesserMoneyMultiplier()));
-            playerStats.changeStat("tPoint", (int)Math.round((double)50.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getMoneyMultiplier()));
-            worldStats.changeStat("wPoint", 110);
-            PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).addExp((double)150.0F);
-            PlayerTitleChecker.killedPlayer(event.getEntity().getKiller());
-            PlayerTitleChecker.gotKilledByPlayer((Player)event.getEntity());
-         } else if (event.getEntity().getType() == EntityType.ENDER_DRAGON) {
-            PlayerTitleChecker.killedDragon(event.getEntity().getKiller());
-         }
-      }
+    public static void Event(org.bukkit.event.entity.EntityDeathEvent event) {
+        if (event.getEntity().getKiller() != null && event.getEntity().getKiller().getType() == EntityType.PLAYER) {
+            PlayerStats playerStats = PlayerStats.getPlayerStats(event.getEntity().getKiller());
+            WorldStats worldStats = WorldStats.getWorldStats(event.getEntity().getWorld());
+            playerStats.changeStat("total_kill_count", 1);
+            playerStats.changeStat("money", (int) Math.round((double) 3.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getLesserMoneyMultiplier()));
+            playerStats.changeStat("tPoint", (int) Math.round((double) 25.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getMoneyMultiplier()));
+            worldStats.changeStat("wPoint", 40);
+            PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).addExp(50.0F);
+            if (event.getEntity().getType() == EntityType.PLAYER) {
+                playerStats.changeStat("money", (int) Math.round((double) 12.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getLesserMoneyMultiplier()));
+                playerStats.changeStat("tPoint", (int) Math.round((double) 50.0F * PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).getMoneyMultiplier()));
+                worldStats.changeStat("wPoint", 110);
+                PlayerLevelHandler.getLevelHandler(event.getEntity().getKiller()).addExp(150.0F);
+                PlayerTitleChecker.killedPlayer(event.getEntity().getKiller());
+                PlayerTitleChecker.gotKilledByPlayer((Player) event.getEntity());
+            } else if (event.getEntity().getType() == EntityType.ENDER_DRAGON) {
+                PlayerTitleChecker.killedDragon(event.getEntity().getKiller());
+            }
+        }
 
-      if (event.getEntity().getType() == EntityType.PLAYER) {
-         PlayerStats.getPlayerStats((Player)event.getEntity()).changeStat("death_count", 1);
-      }
+        if (event.getEntity().getType() == EntityType.PLAYER) {
+            PlayerStats.getPlayerStats((Player) event.getEntity()).changeStat("death_count", 1);
+        }
 
-      WorldStats.getWorldStats(event.getEntity().getWorld()).changeStat("death_count", 1);
-      ServerStats.getServerStats().changeStat("death_count", 1);
-   }
+        WorldStats.getWorldStats(event.getEntity().getWorld()).changeStat("death_count", 1);
+        ServerStats.getServerStats().changeStat("death_count", 1);
+    }
 }
